requirements.txt
flask
pyyaml
requests
docker-py
